Add files via upload
This directory contains sources for our submitted paper for VVIOT2019.

It contains : 
- A source file (MPL_modeling_T1.xlia) for modeling a node and interactions between all nodes in topology T1
- A workflow file (workflow_Update_all_nodes.sew) for executing our heuristics to cover the sequence SEQ1 + FOR1 detailed in the paper. This objective coverage finishes when all nodes in the topology are updated

After executing the workflow, we obtain the result of symbolic execution. An illustration for this result is found in the file /output/basic/symbex_sequence.txt.

PlantUML allows to visualize this file under the form of sequence diagram. For this, you need to download PlantUML here : http://plantuml.com/download